add a -f option to ignore existing lockfiles

+++ serial.c
@@ -27,7 +27,7 @@
 
 int crnl_mapping;		//0 - no mapping, 1 mapping
 struct termios pots;		/* old port termios settings to restore */
-char lockfile[PATH_MAX+1] = "/var/lock/LCK..";
+char *lockfile;
 
 static void init_comm(struct termios *pts)
 {
@@ -102,7 +102,8 @@ static void serial_exit(struct ios_ops *ios)
 	tcsetattr(ios->fd, TCSANOW, &pots);
 	close(ios->fd);
 	free(ios);
-	unlink(lockfile);
+	if (lockfile)
+		unlink(lockfile);
 }
 
 struct ios_ops * serial_init(char *device)
@@ -115,6 +116,9 @@ struct ios_ops * serial_init(char *device)
 
 	ops = malloc(sizeof(*ops));
 	if (!ops)
+		return NULL;
+	lockfile = malloc(PATH_MAX);
+	if (!lockfile)
 		return NULL;
 
 	ops->set_speed = serial_set_speed;
@@ -129,22 +133,33 @@ struct ios_ops * serial_init(char *device)
 	else
 		substring = device;
 
-	strncat(lockfile, substring, PATH_MAX - strlen(lockfile) - 1);
+	sprintf(lockfile, "/var/lock/LCK..%s", substring);
 
 	fd = open(lockfile, O_RDONLY);
-	if (fd >= 0) {
+	if (fd >= 0 && !opt_force) {
 		close(fd);
 		main_usage(3, "lockfile for port exists", device);
 	}
 
+	if (fd >= 0 && opt_force) {
+		close(fd);
+		printf("lockfile for port exists, ignoring\n");
+		unlink(lockfile);
+	}
+
 	fd = open(lockfile, O_RDWR | O_CREAT, 0444);
+	if (fd < 0 && opt_force) {
+		printf("cannot create lockfile. ignoring\n");
+		lockfile = NULL;
+		goto force;
+	}
 	if (fd < 0)
 		main_usage(3, "cannot create lockfile", device);
 	/* Kermit wants binary pid */
 	pid = getpid();
 	write(fd, &pid, sizeof(long));
 	close(fd);
-
+force:
 	/* open the device */
 	fd = open(device, O_RDWR);
 	ops->fd = fd;
